requirements.txt
Flask==3.0.3
docx2pdf==0.2.0
Pillow==10.3.0
pdf2image==1.17.0
weasyprint==60.1
gunicorn==21.2.0
werkzeug